Added booklist feature
In this commit I have added a model booklist
, added a dropdown in viewBook, made a
booklistDAO for sql and added the action in
ViewBookAction. In this it is possible to get
all the books for the customer that is loged in.
It is not implemented a check for the view if the
customer is not loged in. Next step is  to
put a button and make a sql query that add the
book to the customers booklists.