Fixed bug in view book
When adding a book twice to the same
booklist, it crashed because the primary
key consists of the bookid and the listid.
Also removed a bug that information was
removed when adding the book to a booklist.
Have tested with all combinations I have found
and it now works. Also added a message to the user
if the book was added or if the book already
was in the chosen booklist. Woopwoop